# domain/model/src/main/java/co/com/bancolombia/model/response/Cliente.java
package co.com.bancolombia.model.response;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder(toBuilder = true)
public class Cliente {
    private int id;
    private String primerNombre;
    private String segundoNombre;
    private String primerApellido;
    private String segundoApellido;
    private int telefono;
    private String direccion;
    private String ciudad;
    private String correo;
}