# openApi.yaml
openapi: 3.0.0
info:
  title: API de Clientes
  description: API para prueba tecnica BBVA para gestionar clientes
  version: 1.0.0
paths:
  /api/clientes/{tipoDocumento}/{numeroDocumento}:
    get:
      summary: Obtener información de un cliente
      parameters:
        - name: tipoDocumento
          in: path
          required: true
          schema:
            type: string
        - name: numeroDocumento
          in: path
          required: true
          schema:
            type: string
        - name: withAddress
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCliente'
        '400':
          description: Solicitud incorrecta o tipo de documento inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Cliente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/clientes/{numeroDocumento}:
    put:
      summary: Actualizar información de un cliente
      parameters:
        - name: numeroDocumento
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cliente'
      responses:
        '200':
          description: Cliente actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCliente'
        '404':
          description: Cliente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Cliente:
      type: object
      properties:
        primerNombre:
          type: string
        segundoNombre:
          type: string
        primerApellido:
          type: string
        segundoApellido:
          type: string
        telefono:
          type: string
        direccion:
          type: string
        ciudad:
          type: string
        correo:
          type: string
    ApiResponseCliente:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/Cliente'
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string